progress

+++ src/position.rs
@@ -3,6 +3,6 @@ use rust_decimal::prelude::*;
 #[derive(Debug, Clone)]
 pub struct Position {
     pub asset: String,
-    pub lots: i64,
+    pub lots: isize,
     pub cost_basis: Decimal,
 }
